+++ src/models/leaderboard.cairo
@@ -3,7 +3,7 @@
 struct Leaderboard {
     #[key]
     version: u32,
-    high_score: u128,
+    high_score: u32,
     next_version_timestamp: u64,
 }
